Run at/background jobs in their own process group

+++ tests/job-kill/02-loadleveler.t
@@ -15,21 +15,28 @@
 #C: You should have received a copy of the GNU General Public License
 #C: along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #-------------------------------------------------------------------------------
-#C: Test cat-log, localhost
+#C: Test killing of jobs submitted to loadleveler.
 . $(dirname $0)/test_header
 #-------------------------------------------------------------------------------
-N_TESTS=4
+N_TESTS=2
 set_test_number $N_TESTS
-export CYLC_TEST_HOST=$(cylc get-global-config -i '[test battery]remote host')
+export CYLC_TEST_HOST=$( \
+    cylc get-global-config -i '[test battery][directives]loadleveler host')
 if [[ -z $CYLC_TEST_HOST ]]; then
-    skip $N_TESTS '[test battery]remote host: not defined'
+    skip $N_TESTS '[test battery][directives]loadleveler host: not defined'
     exit
 fi
+export CYLC_TEST_DIRECTIVES=$( \
+    cylc get-global-config \
+    -i '[test battery][directives][loadleveler directives]')
 install_suite $TEST_NAME_BASE $TEST_NAME_BASE
 set -eu
-ssh -oBatchMode=yes -oConnectTimeout=5 $CYLC_TEST_HOST \
-    "mkdir -p .cylc/$SUITE_NAME/ && cat >.cylc/$SUITE_NAME/passphrase" \
-    <$TEST_DIR/$SUITE_NAME/passphrase
+if [[ $CYLC_TEST_HOST != 'localhost' ]]; then
+    SSH='ssh -oBatchMode=yes -oConnectTimeout=5'
+    $SSH $CYLC_TEST_HOST \
+        "mkdir -p .cylc/$SUITE_NAME/ && cat >.cylc/$SUITE_NAME/passphrase" \
+        <$TEST_DIR/$SUITE_NAME/passphrase
+fi
 set +eu
 #-------------------------------------------------------------------------------
 TEST_NAME=$TEST_NAME_BASE-validate
@@ -38,15 +45,9 @@ run_ok $TEST_NAME cylc validate $SUITE_NAME
 TEST_NAME=$TEST_NAME_BASE-run
 suite_run_ok $TEST_NAME cylc run --reference-test --debug $SUITE_NAME
 #-------------------------------------------------------------------------------
-TEST_NAME=$TEST_NAME_BASE-2
-cylc cat-log -o $SUITE_NAME echo_2.1 >$TEST_NAME.out
-grep_ok '^echo_2.1$' $TEST_NAME.out
-#-------------------------------------------------------------------------------
-TEST_NAME=$TEST_NAME_BASE-3
-cylc cat-log -o $SUITE_NAME echo_3.1 >$TEST_NAME.out
-grep_ok '^echo_3.1$' $TEST_NAME.out
-#-------------------------------------------------------------------------------
-ssh -oBatchMode=yes -oConnectTimeout=5 $CYLC_TEST_HOST \
-    "rm -rf .cylc/$SUITE_NAME cylc-run/$SUITE_NAME"
+if [[ $CYLC_TEST_HOST != 'localhost' ]]; then
+    $SSH $CYLC_TEST_HOST \
+        "rm -rf .cylc/$SUITE_NAME cylc-run/$SUITE_NAME"
+fi
 purge_suite $SUITE_NAME
 exit
